netsrv -> pkg
